Fix rebase: restore main's versions for config, handlers, experiment_handler
- Replaced config.py, cloud_train_handler.py, experiment_handler.py with
  origin/main versions (DoRA/rsLoRA fields, private methods, inline stage
  runners with parallel eval+loss support)
- Removed orphaned stage runner files (hf_training_stage_runner.py,
  hf_eval_stage_runner.py, hf_loss_stage_runner.py) — main has these
  classes inline in experiment_handler.py
- Fixed test patch targets from stage runner modules to experiment_handler
- Reverted public method names back to private (_apply_training_overrides,
  _load_method_labels, _build_config_display) in tests and pipeline handler
